Added new QA buttons

+++ PlainBasher/Assets/Scripts/Player.cs
@@ -13,17 +13,20 @@ public static class Player {
         }
         set
         {
-			int dmg = value - lives;            
-			if (dmg < 0) OnTakeDamage(dmg);
-			else OnHeal(dmg*-1);
-
-            if (lives > 0 && value <= 0)
+            if (!QADebugging.staticRef.godmode)
             {
-				OnGameOver ();
+			    int dmg = value - lives;            
+			    if (dmg < 0) OnTakeDamage(dmg);
+			    else OnHeal(dmg*-1);
+
+                if (lives > 0 && value <= 0)
+                {
+				    OnGameOver ();
+                }
+			    else if (lives > value && value > 0)
+				    Grid.WipeBoard();
+			    lives = value;
             }
-			else if (lives > value && value > 0)
-				Grid.WipeBoard();
-			lives = value;
         }
     }
 
@@ -47,7 +50,7 @@ public static class Player {
 		Settings.instance.SetPause(true);
 		//TODO: Initialize game end
         guiPlayer.staticref.enabled = false;
-		QADebugging.staticRef.hasLost = true;
+
         guiScore.staticRef.enabled = true;
         
 	}
